+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Vowels.java
@@ -29,7 +29,7 @@ public class Vowels implements ICommand<MessageEvent<Bot>>
 				}
 			}
 
-			Utilities.respond(event, result, false);
+			Utilities.chanMsg(event, result);
 		}
 		else
 			throw new IncorrectCommandExecutionException(getAlias());